Don't send join/part messages from other channels.
When running MinecraftBot on an IRC bouncer to have the same username
for all channels, when a user joins any channel that the bouncer
account is connected to it will print send all the channels they joined
to the Minecraft servers. This commit solves this, and prevents a lot
of potential spam.